WINDUP-1262 - Fix for separating out how windup is ran after the windup server has been started. In tests, windup should not be ran in a separate threaded-job. Also, fix for setting bin/windup to executable prior to running it and preventing a permission exception.
