+++ plugins/org.jboss.tools.windup.ui/META-INF/MANIFEST.MF
@@ -64,4 +64,5 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
 Export-Package: org.jboss.tools.windup.ui,
  org.jboss.tools.windup.ui.internal.explorer,
- org.jboss.tools.windup.ui.internal.services
+ org.jboss.tools.windup.ui.internal.services,
+ org.jboss.tools.windup.ui.util
