+++ plugins/org.jboss.tools.windup.ui/src/org/jboss/tools/windup/ui/internal/messages.properties
@@ -16,6 +16,9 @@ WindupStartingDetail=Windup is being started in server mode.
 WindupStartingError=Error occurred while trying to start Windup in server mode.
 WindupServerError=Windup Server Process Terminated
 
+WindupNotExecutableTitle=Error while running Windup
+WindupNotExecutableInfo=The Windup script is not executable.
+
 showWindupGettingStarted=Show getting started with Windup
 
 showInIssueExplorer=Show Issue in Issue Explorer
